code reorg: break splitstore.go into smaller logical units
